Add server host and username too as env variables.

+++ remote.sh
@@ -2,7 +2,7 @@
 # @Date:   2016-02-15T18:16:05+01:00
 # @Email:  benjamin.burgy@gmail.com
 # @Last modified by:   minidfx
-# @Last modified time: 2016-03-22T20:12:05+01:00
+# @Last modified time: 2016-03-22T20:17:46+01:00
 
 #!/bin/bash
 
@@ -16,6 +16,16 @@ if [ -z $PEBBLE_SDK_PATH ]; then
   exit 2
 fi
 
+if [ -z $USERNAME ]; then
+  echo "USERNAME environment variable is empty or not defined."
+  exit 3
+fi
+
+if [ -z $SERVER_HOST ]; then
+  echo "SERVER_HOST environment variable is empty or not defined."
+  exit 4
+fi
+
 rsync -zrh --delete --exclude ".git/" --exclude "includes/" --progress . couloir42.local:$PROJECT_PATH
 
-ssh minidfx@couloir42.local "cd $PROJECT_PATH;$PEBBLE_SDK_PATH/pebble $@"
+ssh $USERNAME@$SERVER_HOST "cd $PROJECT_PATH;$PEBBLE_SDK_PATH/pebble $@"
